Avoid actually running the engine module in time warp
As it turns out, the patch to ModuleEngines.TimeWarping was causing the
engine to run in time warp, which resulted in twice the amount of fuel
being consumed as it should have been.

This commit adjusts the patch to only replace the call to
DeactivateLoopingFX instead of using a prefix to disable the whole
method.