Keep buffer part resources around as zero-sized inventories
Adding and removing a part resource is quite expensive. It kicks off a
couple of game events that cause a lot of code to run. This results in
stutters whenever we do it. Changing the size of a part resource, in
comparison, is cheap.

This changes the engine buffering system to keep the buffers around as
zero-sized part resources, so that the next time they are needed we can
just start changing their sizes.

Fixes #2